// src/portfolio/dto/project.dto.ts
import { createZodDto } from 'nestjs-zod';
import { z } from 'zod';
import { publicUserSchema } from '../../users/dto/user.dto';

export const projectSchema = z.object({
  id: z.string(),
  title: z.string(),
  description: z.string().optional(),
  links: z.array(z.string()),
  previewUrl: z.string().optional(),
  author: publicUserSchema,
  createdAt: z.iso.datetime(),
  updatedAt: z.iso.datetime(),
});

export class ProjectDto extends createZodDto(projectSchema) {}